7699 review changes
