+++ netbox/virtualization/forms/bulk_import.py
@@ -81,7 +81,7 @@ class ClusterImportForm(NetBoxModelImportForm):
         model = Cluster
         fields = ('name', 'type', 'group', 'status', 'scope_type', 'scope_id', 'tenant', 'description', 'comments', 'tags')
         labels = {
-            'scope_id': 'Scope ID',
+            'scope_id': _('Scope ID'),
         }
